fix saldo

+++ database/migrations/2022_12_20_113038_create_saldos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('saldos', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('saldo');
+            $table->unsignedBigInteger('saldo');
             $table->timestamps();
         });
     }
